Add functionality to change book's shelf
This includes adding Book's shelf changer dropdown-box to each book and also making book's current shelf a default value for the dropdown-box. Selecting different shelf in the dropdown-box will change the book's shelf and book's listing accordingly.

Also, defined a single place to specify all the shelf names, so that if we have to change the shelf name we only have to change it at one place.